# source/api.yaml
openapi: 3.0.3
info:
  version: 0.0.1
  title: ToDo API
paths:
  /api/auth/csrf:
    get:
      operationId: GetCsrfToken
      tags:
        - auth
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CSRFModel"
  /api/auth/signup:
    post:
      operationId: SignUp
      tags:
        - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignUpModel"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserReadModel"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefaultResponseModel"
        "422":
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefaultResponseModel"
  /api/auth/signin:
    post:
      operationId: SignIn
      tags:
        - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignInModel"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserReadModel"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefaultResponseModel"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefaultResponseModel"
  /api/auth/signout:
    get:
      operationId: SignOut
      tags:
        - auth
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefaultResponseModel"
  /api/v1/task:
    post:
      operationId: Create
      tags:
        - v1/task
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskCreateModel"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskReadModel"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefaultResponseModel"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefaultResponseModel"
    put:
      operationId: Update
      tags:
        - v1/task
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdateModel"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskReadModel"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefaultResponseModel"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefaultResponseModel"
  /api/v1/task/done/{id}:
    get:
      operationId: Done
      tags:
        - v1/task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskReadModel"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefaultResponseModel"
  /api/v1/task/{id}:
    get:
      operationId: Get
      tags:
        - v1/task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskReadModel"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefaultResponseModel"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefaultResponseModel"
    delete:
      operationId: Delete
      tags:
        - v1/task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefaultResponseModel"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefaultResponseModel"
  /api/v1/tasks:
    get:
      operationId: GetAll
      tags:
        - v1/task
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskReadModel"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefaultResponseModel"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefaultResponseModel"

components:
  schemas:
    CSRFModel:
      type: object
      properties:
        csrf:
          type: string
      required:
        - csrf
    SignUpModel:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
      required:
        - name
        - email
        - password
    SignInModel:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    UserReadModel:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
      required:
        - id
        - name
        - email
    TaskCreateModel:
      type: object
      properties:
        title:
          type: string
      required:
        - title
    TaskReadModel:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        title:
          type: string
        done_flag:
          type: boolean
      required:
        - id
        - user_id
        - title
        - done_flag
    TaskUpdateModel:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
      required:
        - id
        - title
    DefaultResponseModel:
      type: object
      properties:
        message:
          type: string
      required:
        - message